Split wren.py and basic cleanup.
